+++ packages/auto-complete-element/src/autocomplete.ts
@@ -11,6 +11,7 @@ export default class Autocomplete {
   element: AutoCompleteElement;
   input: HTMLInputElement;
   list: HTMLElement;
+  resetButton: HTMLElement | null;
   isMultiple: boolean;
   combobox: Combobox;
   listObserver: MutationObserver;
@@ -27,15 +28,25 @@ export default class Autocomplete {
     this.input.setAttribute('spellcheck', 'false');
     this.input.setAttribute('autocomplete', 'off');
 
+    // Reset button
+    this.resetButton = this.element.querySelector('[data-autocomplete-reset]');
+    if (this.resetButton && !this.resetButton.hasAttribute('aria-label')) {
+      this.resetButton.setAttribute('aria-label', 'reset autocomplete');
+    }
+
     this.onFocus = this.onFocus.bind(this);
+    this.onPointerDown = this.onPointerDown.bind(this);
     this.onKeydown = this.onKeydown.bind(this);
     this.onCommit = this.onCommit.bind(this);
     this.onInput = debounce(this.onInput.bind(this), 300);
+    this.handleReset = this.handleReset.bind(this);
 
     this.input.addEventListener('focus', this.onFocus);
+    this.input.addEventListener('pointerdown', this.onPointerDown); // We use `pointerdown` instead of `click` to simulate the `focus` event
     this.input.addEventListener('keydown', this.onKeydown);
     this.input.addEventListener('input', this.onInput);
     this.list.addEventListener('combobox:commit', this.onCommit);
+    this.resetButton?.addEventListener('click', this.handleReset);
 
     this.listObserver = new MutationObserver(this.onListToggle.bind(this));
     this.listObserver.observe(this.list, { attributes: true, attributeFilter: ['hidden'] });
@@ -44,9 +55,11 @@ export default class Autocomplete {
 
   destroy() {
     this.input.removeEventListener('focus', this.onFocus);
+    this.input.removeEventListener('pointerdown', this.onPointerDown);
     this.input.removeEventListener('keydown', this.onKeydown);
     this.input.removeEventListener('input', this.onInput);
     this.list.removeEventListener('combobox:commit', this.onCommit);
+    this.resetButton?.removeEventListener('click', this.handleReset);
 
     this.listObserver.disconnect();
   }
@@ -64,9 +77,14 @@ export default class Autocomplete {
   onFocus() {
     if (!this.list.hidden) return;
 
-    this.list.hidden = false;
-    this.combobox.options.forEach(filterOptions('', { matching: AUTOCOMPLETE_VALUE_ATTR }));
-    activateFirstOption(this);
+    this.openAndInitializeList();
+  }
+
+  onPointerDown() {
+    if (!this.list.hidden) return;
+    if (document.activeElement !== this.input) return; // If it's not already active, then `onFocus` logic will apply
+
+    this.openAndInitializeList();
   }
 
   onKeydown(event: KeyboardEvent) {
@@ -80,9 +98,7 @@ export default class Autocomplete {
         break;
       case 'ArrowDown':
         if (event.altKey && this.list.hidden) {
-          this.list.hidden = false;
-          this.combobox.options.forEach(filterOptions('', { matching: AUTOCOMPLETE_VALUE_ATTR }));
-          activateFirstOption(this);
+          this.openAndInitializeList();
           event.preventDefault();
           event.stopPropagation();
         }
@@ -134,6 +150,30 @@ export default class Autocomplete {
     if (this.list.contains(event.target as HTMLElement)) return;
 
     this.list.hidden = true;
+  }
+
+  async handleReset(event: Event) {
+    event.preventDefault();
+
+    for (const option of this.combobox.options.filter(selected)) {
+      option.setAttribute('aria-selected', 'false');
+    }
+    syncSelection(this);
+    this.input.focus();
+
+    await nextTick();
+    if (!this.list.hidden) {
+      this.list.hidden = true;
+    }
+
+    // Should fire after closing the list
+    this.list.dispatchEvent(new CustomEvent('auto-complete:reset', { bubbles: true }));
+  }
+
+  openAndInitializeList() {
+    this.list.hidden = false;
+    this.combobox.options.forEach(filterOptions('', { matching: AUTOCOMPLETE_VALUE_ATTR }));
+    activateFirstOption(this);
   }
 
   set inputValue(value: string) {